+++ app/routes/components/ThemeProvider/changelog.codegen.ts
@@ -8,4 +8,6 @@ export const changelog = ''
   + '- 2025-09-20: Add useTheme hook and refactor ThemeProvider to make it cleaner.\n'
   + '- 2025-09-23: Add tokens to the theme with prefix `inputBase` for input base\n'
   + '  components.\n'
+  + '- 2025-09-29: Add tokens to the theme with prefix `inputBase` for size `xs` and\n'
+  + '  `xl`.\n'
   ;