+++ src/components/ThemeProvider/index.tsx
@@ -1,3 +1,3 @@
-// ThemeProvider from copy-ui @ 2025-09-23
+// ThemeProvider from copy-ui @ 2025-09-29
 
 export { ThemeProvider, ThemeContext, useTheme } from "./ThemeProvider";